Избраное?

+++ backend/foodgram/api/views.py
@@ -11,12 +11,11 @@ from rest_framework.permissions import IsAuthenticatedOrReadOnly
 from rest_framework.response import Response
 from users.models import Follow, User
 
-from .filters import IngredientFilter
+from .filters import IngredientFilter, RecipeFilter
 from .pagintation import CustomPagination
 from .permissions import IsAdminAuthorOrReadOnly, IsAdminOrReadOnly
 from .serializers import (CreateRecipesSerializer, FavoriteSerializer,
-                          FollowSerializer, InfoFollowSerializer,
-                          IngredientsSerializer, RecipesSerializer,
+                          FollowSerializer, IngredientsSerializer, RecipesSerializer,
                           ShoppingCartSerializer, TagSerializer,
                           UsersSerializer)
 
@@ -41,44 +40,42 @@ class CustomUserViewSet(UserViewSet):
     queryset = User.objects.all()
     serializer_class = UsersSerializer
     permission_classes = [IsAuthenticatedOrReadOnly]
+    pagination_class = CustomPagination
 
-    @action(['get'], detail=False)
+    @action(detail=False)
     def subscriptions(self, request):
-        subscriptions_list = self.paginate_queryset(
-            User.objects.filter(following__user=request.user)
-        )
-        serializer = InfoFollowSerializer(
-            subscriptions_list, many=True, context={
-                'request': request
-            }
-        )
+        user = request.user
+        follows = User.objects.filter(following__user=user)
+        page = self.paginate_queryset(follows)
+        serializer = FollowSerializer(
+            page, many=True,
+            context={'request': request})
         return self.get_paginated_response(serializer.data)
 
-    @action(['POST', 'DELETE'], detail=True)
-    def subscribe(self, request, **kwargs):
+    @action(methods=['POST', 'DELETE'],
+            detail=True, )
+    def subscribe(self, request, id):
         user = request.user
-        author_id = self.kwargs.get('id')
-        author = get_object_or_404(User, id=author_id)
+        author = get_object_or_404(User, id=id)
+        subscription = Follow.objects.filter(
+            user=user, author=author)
 
         if request.method == 'POST':
-            serializer = FollowSerializer(
-                author,
-                data=request.data,
-                context={'request': request}
-            )
-            serializer.is_valid(raise_exception=True)
-            serializer.save()
-            return Response(
-                serializer.data, status=status.HTTP_201_CREATED
-            )
-        elif request.method == 'DELETE':
-            subscription = get_object_or_404(
-                Follow,
-                author=author,
-                user=user
-            )
-            subscription.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
+            if subscription.exists():
+                return Response({'error': 'Вы уже подписаны'},
+                                status=status.HTTP_400_BAD_REQUEST)
+            if user == author:
+                return Response({'error': 'Невозможно подписаться на себя'},
+                                status=status.HTTP_400_BAD_REQUEST)
+            serializer = FollowSerializer(author, context={'request': request})
+            Follow.objects.create(user=user, author=author)
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        if request.method == 'DELETE':
+            if subscription.exists():
+                subscription.delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            return Response({'error': 'Вы не подписаны на этого пользователя'},
+                            status=status.HTTP_400_BAD_REQUEST)
 
 
 class RecipesViewSet(viewsets.ModelViewSet):
@@ -87,6 +84,8 @@ class RecipesViewSet(viewsets.ModelViewSet):
     """
     queryset = Recipe.objects.all()
     serializer_class = RecipesSerializer
+    filter_backends = (DjangoFilterBackend,)
+    filterset_class = RecipeFilter
     permission_classes = (IsAdminAuthorOrReadOnly,)
     pagination_class = CustomPagination
